+++ Realm/RLMSyncUser.h
@@ -47,7 +47,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @interface RLMSyncUser : NSObject
 
-/** 
+/**
  An array of all valid, logged-in users.
  */
 + (NSArray<RLMSyncUser *> *)allUsers NS_REFINED_FOR_SWIFT;